data: Write symbolic refs
We have one last step before we can finally implement branching. We will add
code to `update_ref` that can write a symbolic value.

Note: In a previous change we implemented the ability to write to a symbolic
ref, but we could only write **OIDs** to it. Now we implemented the ability to
write a **symbolic value** to a ref (and it doesn't matter if the ref that we
write is itself symbolic or non-symbolic). These two sound similar and
confusing, but make sure that you pause and understand why these are different
things.

%% METADATA
Id: write-symbolic-refs