diff: Compare working tree to a commit
Let's implement another useful command that will make use of *diff* - a command
that will display what was changed in the working directory since the last
commit. This is useful if you don't remember exactly what you have changed or if
you want to make sure that you're not introducing any unintended changes to a
commit by mistake. The command will be called `diff`, to mirror Git's name.

Our implementation of `diff` will use `diff.diff_trees()` that we created
earlier for `show`. This time, we will ask `diff_trees()` to compare the
"working tree" with the tree of some commit. The "working tree" is a dictionary
that describes the files in the working directory.

We will implement `get_working_tree()`, which will walk over all files in the
working directory, put them in the object database and create a dict that holds
all the OIDs. This dictionary will represent a "tree" without actually writing a
tree object.

(It may seem strange to put all files in the object database each time just for
calculating a diff, but it will be needed later anyway, when we implement
something called the "index".)

Then let's create a CLI command `diff` which will just compare the current
"working tree" to a commit tree (by default to HEAD, but it can be any commit we
want).

%% METADATA
Id: diff-working-tree