merge: Three-way merge
Next, let's improve the quality of our merges by using a technique called
**three-way merge**. Before we talk about three-way merge, let's talk about two-way
merge, which is the technique we've used so far.

Two-way merge means that we look at two versions of the same file and we try to
merge them into one. Sometimes it works well. For example, if we have:

- Version A:

      def be_a_cat ():
          print ("Meow")
          while True:
                print ("Purr")
          return True

      def be_a_dog ():
          print ("Bark!")
          return False

- Version B:

      def be_a_cat ():
          print ("Meow")
          return True

      def be_a_dog ():
          print ("Bark!")
          print ("Wiggle tail")
          return False

It's pretty easy to merge them together, because it's clear from the context
which are the common parts and which parts were added in each version.

- A and B merged:

      def be_a_cat ():
          print ("Meow")
          while True:
                print ("Purr")
          return True

      def be_a_dog ():
          print ("Bark!")
          print ("Wiggle tail")
          return False

And that's exactly what our `diff` CLI command would do. But let's see an
example where two-way merge doesn't work well:

- Version A:

      def be_a_cat ():
          print ("Sleep")
          return True

      def be_a_dog ():
          print ("Bark!")
          return False

- Version B:

      def be_a_cat ():
          print ("Meow")
          return True

      def be_a_dog ():
          print ("Eat homework")
          return False

How should we merge them? We have a conflict in each print statement. It means
that the best we can do is leave both changes and let the user decide:

- A and B merged:

      def be_a_cat ():
      <<<<<<< Version A
          print ("Sleep")
      =======
          print ("Meow")
      >>>>>>> Version B
          return True

      def be_a_dog ():
      <<<<<<< Version A
          print ("Bark!")
      =======
          print ("Eat homework")
      >>>>>>> Version B
          return False

Now assume that version A and version B are the same file but from different
branches. For example like this:

                    version A
                    v
    o---o---o---o---o
         \
          --o---o
                ^
                version B

There's one more interesting piece of information that we didn't use during the
two-way merge: The file might be present in their common ancestor and it might
give us useful context about the merge. For example, assuming we know this:

- Version A:

      def be_a_cat ():
          print ("Sleep")
          return True

      def be_a_dog ():
          print ("Bark!")
          return False

- Version B:

      def be_a_cat ():
          print ("Meow")
          return True

      def be_a_dog ():
          print ("Eat homework")
          return False

- Common ancestor:

      def be_a_cat ():
          print ("Meow")
          return True

      def be_a_dog ():
          print ("Bark!")
          return False

We can actually understand that version A changed only `be_a_cat`'s print
statement while version B changed only `be_a_dog`'s print statement. It means
that each version changed unrelated pieces of the code and we can merge the
safely into:

- A and B merged using common ancestor:

      def be_a_cat ():
          print ("Sleep")
          return True

      def be_a_dog ():
          print ("Eat homework")
          return False

This is called a **three-way merge** - an algorithm that merges two files using
their common ancestor as a guide. It improves merge quality
dramatically (meaning less conflicts on each merge). If you'd like to know more,
please search online :).

We would like to use three-way merges in ugit. For that we're going to use
`get_merge_base` that we've implemented earlier and modify `diff.merge_trees` to
accept a common ancestor (called a "base").

We don't need a lot of logic changes to make it happen. Our `compare_trees`
function can compare an arbitrary number of trees, so we just add the base tree
there as well. Then `merge_blobs` will call `diff3` CLI command which performs
three-way merges.

You can try playing with the resulting code to see that the merge results are
way way better.

%% METADATA
Id: merge-three-way