read-tree: Read into index instead of working directory
For symmetry, we need to populate the index when reading a tree as well.

During commit, the data flows from the working directory to the index to the
commit, so when reading a tree (during checkout or merge) the data will flow
from the tree to the index to the working directory.

Without this change, when the user will checkout a different commit the state of
the index will still reflect the files as they were during the old commit. But
now it will be correctly updated.

%% METADATA
Id: read-tree-to-index